requirements.txt
python-telegram-bot==13.15
requests
pybit==2.3.0
